requirements.txt
asgiref==3.7.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
dj-database-url==2.1.0
Django==4.2.5
django-cors-headers==4.2.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
google-api-core==2.11.1
google-api-python-client==2.100.0
google-auth==2.23.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.60.0
gunicorn==21.2.0
httplib2==0.22.0
idna==3.4
jsonfield==3.1.0
oauthlib==3.2.2
openpyxl==3.1.2
packaging==23.2
protobuf==4.24.3
psycopg2==2.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyJWT==2.8.0
pyparsing==3.1.1
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
whitenoise==6.5.0
